nvme: Add FENCING nvme controller state
Add NVME_CTRL_FENCING as a new controller state to be used when impacted
controller is being fenced. A LIVE controller enters fencing state when
an IO error is encountered. While in fencing state, inflight IOs will not
be canceled if they timeout. While in fencing state a controller can not
be reset or deleted. This is intentional because reset or delete will
result in canceling inflight IOs. After fencing finishes, the controller
transition to RESETTING state where reset codepath takes care of queue
teardown and inflight requests cancelation.

Note, there is no transition from FENCING to RESETTING added to
nvme_change_ctrl_state(). The reason is that user should not be allowed
to reset or delete a controller that is being recovered.

Add NVME_CTRL_FENCED controller flag. This flag is set on a controller
about to schedule delayed work for time based recovery.

Signed-off-by: Mohamed Khalfella <mkhalfella@purestorage.com>